+++ coincube-gui/src/app/view/nav/tertiary.rs
@@ -8,6 +8,7 @@
 //! Renders nothing when the active route has no third level — callers
 //! check [`items_for`] for that and omit the rail entirely.
 
+use super::items::{render_item_row, RAIL_ITEM_HEIGHT};
 use super::{NavContext, SubItem};
 use crate::app::{
     menu::{
@@ -17,24 +18,16 @@ use crate::app::{
     view::Message,
 };
 use coincube_ui::{
-    color,
     icon::{
         bitcoin_icon, chat_icon, graph_icon, home_icon, lightning_icon, plus_icon, receipt_icon,
         settings_icon, tooltip_icon, wallet_icon, wrench_icon,
     },
     theme,
-    widget::{Button, Column, Element, Row},
+    widget::{Column, Element},
 };
-use iced::{
-    widget::{column, container, row, Space},
-    Alignment, Length,
-};
+use iced::{widget::container, Length};
 
 pub const RAIL_WIDTH: f32 = 72.0;
-const ITEM_HEIGHT: f32 = 64.0;
-const HIGHLIGHT_WIDTH: f32 = 5.0;
-const ICON_SIZE: f32 = 22.0;
-const LABEL_SIZE: f32 = 10.0;
 
 /// Returns the tertiary-rail items for `menu`, or `None` when the route
 /// has no third level (and the rail should be hidden).
@@ -54,10 +47,10 @@ pub fn rail<'a>(menu: &Menu, ctx: &NavContext<'a>) -> Option<Element<'a, Message
     // background only extends behind the items themselves. No trailing
     // fill — when used as an overlay the bottom of the rail needs to
     // stay click-through so content beneath it stays interactive.
-    let items_height = items.len() as f32 * ITEM_HEIGHT;
+    let items_height = items.len() as f32 * RAIL_ITEM_HEIGHT;
     let mut items_col: Column<Message> = Column::new().spacing(0).width(Length::Fill);
     for item in items {
-        items_col = items_col.push(item_row(menu, &item));
+        items_col = items_col.push(render_item_row(menu, &item, RAIL_WIDTH));
     }
 
     Some(
@@ -67,61 +60,6 @@ pub fn rail<'a>(menu: &Menu, ctx: &NavContext<'a>) -> Option<Element<'a, Message
             .style(theme::container::sidebar_tertiary)
             .into(),
     )
-}
-
-fn item_row<'a>(menu: &Menu, item: &SubItem) -> Element<'a, Message> {
-    let active = (item.matches)(menu);
-    let icon = (item.icon)().size(ICON_SIZE);
-
-    let body: Column<Message> = column![
-        icon,
-        iced::widget::text(item.label)
-            .size(LABEL_SIZE)
-            .align_x(Alignment::Center),
-    ]
-    .spacing(4)
-    .align_x(Alignment::Center)
-    .width(Length::Fill);
-
-    let on_press = if active {
-        Message::Reload
-    } else {
-        Message::Menu(item.route.clone())
-    };
-
-    let button: Button<Message> = Button::new(
-        container(body)
-            .padding([8, 0])
-            .width(Length::Fill)
-            .center_x(Length::Fill),
-    )
-    .width(Length::Fill)
-    .height(Length::Fixed(ITEM_HEIGHT))
-    .style(if active {
-        theme::button::rail_active
-    } else {
-        theme::button::rail
-    })
-    .on_press(on_press);
-
-    // Orange strip on the trailing (right) edge for the active item —
-    // same treatment as the secondary rail.
-    let highlight: Element<'a, Message> = if active {
-        container(Space::new().width(Length::Fixed(HIGHLIGHT_WIDTH)))
-            .height(Length::Fixed(ITEM_HEIGHT))
-            .style(theme::container::custom(color::ORANGE))
-            .into()
-    } else {
-        container(
-            Space::new()
-                .width(Length::Fixed(HIGHLIGHT_WIDTH))
-                .height(Length::Fixed(ITEM_HEIGHT)),
-        )
-        .into()
-    };
-
-    let r: Row<Message> = row![button, highlight].width(Length::Fixed(RAIL_WIDTH));
-    r.into()
 }
 
 fn home_settings_items() -> Vec<SubItem> {
